# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rento Vroom API",
    "version": "1.0.0",
    "description": "REST API for the Rento Vroom website and future apps. Every error response is `{ error: { code, message, fields? } }`. Requests that change data from a browser must come from the website's own origin."
  },
  "servers": [
    {
      "url": "https://api.rentovroom.com/api/v1",
      "description": "Production"
    },
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "rv_access",
        "description": "The 15-minute access token, set by the sign-in routes (the website)"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "The same access token in an Authorization header (mobile apps)"
      }
    },
    "schemas": {
...
      "AdminOverview": {
        "type": "object",
        "properties": {
          "metrics": {
            "type": "object",
            "properties": {
              "totalUsers": {
                "type": "integer"
              },
              "activeHosts": {
                "type": "integer"
              },
              "staffMembers": {
                "type": "integer"
              },
              "suspendedUsers": {
                "type": "integer"
              },
              "activeVehicles": {
                "type": [
                  "integer",
                  "null"
                ]
              },
              "upcomingBookings": {
                "type": [
                  "integer",
                  "null"
                ]
              },
              "bookingRevenueCents": {
                "type": [
                  "integer",
                  "null"
                ]
              },
              "pendingVerifications": {
                "type": [
                  "integer",
                  "null"
                ]
              }
            },
            "required": [
              "totalUsers",
              "activeHosts",
              "staffMembers",
              "suspendedUsers",
              "activeVehicles",
              "upcomingBookings",
              "bookingRevenueCents",
              "pendingVerifications"
            ]
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "metrics",
          "generatedAt"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/admin/overview": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "KPI figures for the staff portal",
        "description": "Admin and Support only.",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The overview figures",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOverview"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session has ended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but this account can't do this (or the request came from an untrusted origin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/staff/{id}/mfa/reset": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Reset a staff member's lost authenticator apps",
        "description": "Admin only, and not for your own account. Removes all their authenticator apps and signs them out everywhere; they sign in with only their password and can set up a new app in Settings. Written to the audit log.",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "The staff member’s user id"
            },
            "required": true,
            "description": "The staff member’s user id",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Reset"
          },
          "401": {
            "description": "Not signed in, or the session has ended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but this account can't do this (or the request came from an untrusted origin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/payments/test": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Start a NZ$1 test payment in the Stripe sandbox",
        "description": "Admin only. Checks the Stripe keys, Apple Pay, Google Pay and the webhook before the booking flow exists. Refused (409 LIVE_MODE) with live keys, and 503 until the keys are set.",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "The payment, ready for Stripe.js to confirm",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestPayment"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session has ended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but this account can't do this (or the request came from an untrusted origin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with existing data, e.g. the email address already has an account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A service this needs isn't set up or available yet, e.g. payments before the Stripe keys are set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/payments/test/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "How a test payment went",
        "description": "Admin only. Its status, how it was paid (card, Apple Pay, Google Pay) and whether its webhook arrived.",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "The PaymentIntent id (pi_…)"
            },
            "required": true,
            "description": "The PaymentIntent id (pi_…)",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "The test payment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestPaymentStatus"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session has ended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but this account can't do this (or the request came from an untrusted origin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A service this needs isn't set up or available yet, e.g. payments before the Stripe keys are set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/vehicles/{slug}": {
      "get": {
        "tags": [
          "Vehicles"
        ],
        "summary": "A live listing",
        "description": "Public. Approved photos, specs, rego and WOF status, policies, delivery options and the Host. Never the plate or the exact address. 404 for anything not live.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "slug",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "The listing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/host-applications": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Host applications",
        "description": "Oldest first. `status` is APPLIED by default.",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "APPLIED",
                "APPROVED",
                "REJECTED",
                "SUSPENDED"
              ]
            },
            "required": false,
            "name": "status",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Applications",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostApplications"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session has ended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but this account can't do this (or the request came from an untrusted origin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/host-applications/{userId}/approve": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Approve a Host application",
        "description": "409 EMAIL_NOT_VERIFIED until the applicant has confirmed their email. Emails the Host.",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "userId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewNotes"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The new status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostDecision"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session has ended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but this account can't do this (or the request came from an untrusted origin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with existing data, e.g. the email address already has an account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/host-applications/{userId}/reject": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Reject a Host application",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "userId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequiredReviewNotes"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The new status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input. `error.fields` has one message per invalid field.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session has ended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but this account can't do this (or the request came from an untrusted origin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/vehicles": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "The listing review queue",
        "description": "Listings under review, and live listings with new photos or documents waiting.",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The queue",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewQueue"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session has ended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but this account can't do this (or the request came from an untrusted origin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/vehicles/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "A listing to review, with its Host",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "The car’s id"
            },
            "required": true,
            "description": "The car’s id",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "The listing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminVehicle"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session has ended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but this account can't do this (or the request came from an untrusted origin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/vehicles/{id}/approve": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Approve a listing (and its pending photos and documents)",
        "description": "409 HOST_NOT_APPROVED until the Host’s application is approved.",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "The car’s id"
            },
            "required": true,
            "description": "The car’s id",
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewNotes"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The listing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostVehicleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input. `error.fields` has one message per invalid field.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session has ended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but this account can't do this (or the request came from an untrusted origin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with existing data, e.g. the email address already has an account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/vehicles/{id}/request-changes": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Send a listing back to the Host for changes",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "The car’s id"
            },
            "required": true,
            "description": "The car’s id",
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequiredReviewNotes"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The listing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostVehicleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input. `error.fields` has one message per invalid field.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session has ended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but this account can't do this (or the request came from an untrusted origin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with existing data, e.g. the email address already has an account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/vehicles/{id}/calendar": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "A car’s calendar",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "The car’s id"
            },
            "required": true,
            "description": "The car’s id",
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "from",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "to",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "The calendar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostCalendar"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session has ended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but this account can't do this (or the request came from an untrusted origin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}